requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
PyMySQL==1.1.1
SQLAlchemy==2.0.32
requests==2.32.3
python-dotenv==1.0.1
python-json-logger==2.0.7
python-dateutil==2.9.0.post0
